Sort the corrections based on their weights.
Weights are primarily based on the edit distance but also
on the type of edits. For now, inserts are considered more
important than anything else.

Also, it is very rare for someone to type the first character
of a spelling incorrectly, therefore punish the weights for
corrections where edits are happening on position 1.