Refactor:
Remove functions: spell, get_corrections
Move function: is_known_word to spellutils.c

spellutils.c is the place to keep functions which are supposed to
be used internally by other files but not to be exposed as a public
API, in case libspell does become a library.

Get rid of spell.c, not needed anymore

This helps reduce the size of most of the binaries drastically as
now they don't depend on websters.c at all. websters.c is now needed
only when generating ngrams or soundex codes.