Remove the ngram parameter from the API spell_get_suggestions.
It was making the API complicated without adding any value. For it be
useful the caller would need to parse and pass the correct number of words
and then we would have to parse the input again. This is confusing and a better
API design needs to be thought of, until then good bye to this argument.