Revert "Make changes in trie to store void * values."
Using tries for storing the soundex data regresses the memory and
run time performance. I guess because of the shorter length of soundex
codes and many repeated characters across the codes, trie is not suited
for it. No point in wasting memory for storing a pointer when we only
need to store the word frequencies.

This reverts commit 150e9969c57635155b57eff38131496f940f5d92.